Extract common renderings
